Keep profile package filters out of the runtime worker
Search after a filter pill now re-renders the shell so hidden inputs match the URL. Owner-only query params stay off the guest search form. Filter SQL lives in an origin-only module so kody-runtime stays under its startup budget.
